// src/app/models/examination/palliative-factors.ts
export enum PalliativeFactors {
  NONE = "None",
  RELIEVED_BY_NITROGLYCERIN_WITHIN_5_MIN = "Relieved by nitroglycerin within 5 min",
  RELIEVED_BY_NITROGLYCERIN_AFTER_MORE_THAN_5_MIN = "Relieved by nitroglycerin after more than 5 min",
  ANTACIDS = "Antacids",
  ANALGESIA_EXCEPT_MORPHINE = "Analgesia except morphine",
  MORPHINE = "Morphine",
}