// src/app/models/examination/history-of-similiar-pain.ts
export enum HistoryOfSimiliarPain {
  PRIOR_CHEST_PAIN_OF_THIS_TYPE = "Prior chest pain of this type",
  PHYSICIAN_CONSULTED_FOR_PRIOR_PAIN = "Physician consulted for prior pain",
  PRIOR_PAIN_RELATED_TO_HEART = "Prior pain related to heart",
  PRIOR_PAIN_DUE_TO_MI = "Prior pain due to MI",
  PRIOR_PAIN_DUE_TO_ANGINA_PRECTORIS = "Prior pain due to angina prectoris",
}