// src/app/models/examination/past-medical-history.ts
export enum PastMedicalHistory {
  PRIOR_MI = "Prior MI",
  PRIOR_ANGINA_PRECTORIS = "Prior angina prectoris",
  PRIOR_ATYPICAL_CHEST_PAIN = "Prior atypical chest pain",
  CONGESTIVE_HEART_FAILURE = "Congestive heart failure",
  PERIPHERAL_VASCULAR_DISEASE = "Peripheral vascular disease",
  HYPERTENSION = "Hypertension",
  DIABETES = "Diabetes",
  SMOKER = "Smoker",
}