// src/app/models/examination/physical-examinations.ts
export enum PhysicalExaminations {
  RALES = "RALES",
  CYANOIS = "Cyanois",
  PALLOR = "Pallor",
  SYSTOLIC_MURMUR = "Systolic murmur",
  DIASTOLIC_MURMUR = "Diastolic murmur",
  OEDEMA = "Oedema",
  S3_GALLOP = "S3 gallop",
  S4_GALLOP = "S4 gallop",
  CHEST_WALL_TENDERNESS = "Chest wall tenderness",
  DIAPHORESIS = "Diaphoresis",
}